[Sprint 2 ]  Elasticsearch — Indexing, Sync & Full-text Search

+++ src/products/products.service.ts
@@ -1,19 +1,31 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import { ConflictException, Injectable, NotFoundException } from "@nestjs/common";
 import { CreateProductDto } from "./dto/create-product.dto";
 import { PrismaService } from "../prisma/prisma.service"; 
 import { UpdateProductDto } from "./dto/update-product.dto";
 import { RedisService } from "../redis/redis.service";
+import { SearchService } from "../search/search.service";
 
 @Injectable()
 export class ProductsService {
 
-  constructor(private prisma: PrismaService, private redisService: RedisService) {}
+  constructor(
+    private prisma: PrismaService,
+    private redisService: RedisService,
+    private searchService: SearchService
+  ) {}
 
   async createProduct( dto: CreateProductDto ): Promise<{ product: any }> {
+    const existingProduct = await this.prisma.product.findUnique({
+      where: { sku: dto.sku },
+    });
+    if (existingProduct) {
+      throw new ConflictException(`Product with SKU ${dto.sku} already exists`);
+    }
     const product = await this.prisma.product.create({
       data: dto
     });
     await this.redisService.deletePattern(`products:all`);
+    await this.searchService.indexProduct(product);
     return { product };
   }
 
@@ -60,6 +72,7 @@ export class ProductsService {
     });
     await this.redisService.delX(`product:${updatedProduct.id}`);
     await this.redisService.deletePattern(`products:all`); // Invalidate the all products cache
+    await this.searchService.indexProduct(updatedProduct);
     return { product: updatedProduct };
   }
 
@@ -79,6 +92,7 @@ export class ProductsService {
   });
 
   await this.redisService.deletePattern(`products:all`); // Invalidate the all products cache
+  await this.searchService.removeProduct(id);
   return { message: `Product with id ${id} has been deactivated` };
 }
 }
